requirements.txt
asgiref==3.2.3
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-ajax-selects==1.9.1
django-allauth==0.41.0
django-appconf==1.0.4
django-autocomplete-light==3.5.1
django-crispy-forms==1.9.1
django-formtools==2.2
django-maintenancemode==0.11.3
django-phonenumber-field==4.0.0
django-widget-tweaks==1.4.5
httplib2==0.9.2
idna==2.9
IPy==1.0
oauthlib==3.1.0
pendulum==2.0.5
phonenumbers==8.12.4
Pillow==7.0.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
pytzdata==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.9
